Refactoring and redesign cluster
git-svn-id: https://svn.apache.org/repos/asf/tomcat/container/catalina/trunk@303857 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/cluster/src/share/org/apache/catalina/cluster/tcp/DataSender.java
@@ -231,8 +231,8 @@ public class DataSender implements IDataSender {
     /**
      * @return Returns the avg totalBytes/nrOfRequests.
      */
-    public double getAvgMessageSize() {
-        return ((double)totalBytes) / nrOfRequests;
+    public long getAvgMessageSize() {
+        return totalBytes / nrOfRequests;
     }
 
     /**
